requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
numpy==1.25.2
psycopg2==2.9.6
psycopg2-binary==2.9.7
pytz==2023.3
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
setuptools==65.5.0
shapely==2.0.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
